requirements_a.txt
bitsandbytes==0.47.0
datasets==4.0.0
huggingface_hub==0.34.4
inflect==7.5.0
ipython==9.4.0
lxml==6.0.1
peft==0.17.1
python-dotenv==1.1.1
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.55.3
